+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 2.12.0-beta.1 (Unreleased)
 
+### Features Added
+
+- Supported option of `filter` and `expand` for list instance of virtual machines by `VirtualMachineScaleSetVMs.list`.
+- Changed to include the instance view of the virtual machine, when getting the virtual machine by `VirtualMachineScaleSetVMs.getInstance`.
+
 ### Other Changes
 
 #### Dependency Updates